requirements.txt
pymongo
st-annotated-text
pyyaml
pandas